requirements.txt
gradio
tensorflow==2.18.0
tf-keras
transformers==4.52.4
tensorflow-text==2.18.0
numpy
pandas
nltk